update: - extracted SessionPresenters component from modal to its own component so it can be reused in mobile view
