+++ src/components/grid/gridRowContent.js
@@ -5,21 +5,22 @@ import { SessionTime } from 'SVComponents/sessionTime/sessionTime'
 import { useTheme } from '@keg-hub/re-theme'
 import PropTypes from 'prop-types'
 import { SessionLink } from 'SVComponents/sessionLink'
+import {EvfTextToggle} from 'SVComponents/textToggle'
 import {
   View,
   Text,
   Drawer,
   Touchable,
-  TextToggle,
 } from '@keg-hub/keg-components'
 import { useSessionLocation } from 'SVHooks/models'
 import { BookingButton } from 'SVComponents/button'
+import {SessionPresenters} from 'SVComponents/sessionDetails'
 
 /**
  * The content of a grid item when displayed as a row (<= 480px width)
  * @param {Object} props
  * @param {Array} props.labels - the array of label model objects
- * @param {Object} props.session - the session model object
+ * @param {import('SVModels/session').Session} props.session - the session model object
  * @param {Object} props.labelStyles - styles for individual labels
  * @param {boolean} props.militaryTime - if true, use military time for dates
  */
@@ -68,7 +69,10 @@ export const GridRowContent = props => {
 }
 
 /**
- * @todo: to be completed in https://jira.simpleviewtools.com/browse/ZEN-391
+ * 
+ * @param {object} props 
+ * @param {import('SVModels/session').Session} props.session
+ * @param {object} props.styles
  */
 const DrawerContent = ({ session, styles }) => {
   return (
@@ -77,9 +81,11 @@ const DrawerContent = ({ session, styles }) => {
         session={session}
         styles={styles?.bookingButton}
       />
-      <TextToggle
+      <SessionPresenters 
+        session={session}
+      />
+      <EvfTextToggle
         text={session.summary}
-        styles={styles?.toggleText}
       />
     </View>
   )
