+++ src/components/sessionDetails/index.js
@@ -1,0 +1,1 @@
+export * from './sessionPresenters'
