+++ src/components/tapIndex.js
@@ -7,5 +7,7 @@ export * from './hocs/tapIndex'
 export * from './button'
 export * from './sessionLink'
 export * from './loading'
+export * from './textToggle'
+export * from './sessionDetails'
 export * from 'SVComponents/kegComponents'
 export * from 'SVComponents/sessions'
