+++ src/theme/components/sessionDetails/index.js
@@ -1,0 +1,5 @@
+import {sessionPresenters} from './sessionPresenters'
+
+export const sessionDetails = {
+  sessionPresenters,
+}
