+++ src/theme/components/sessionDetails/sessionPresenters.js
@@ -1,0 +1,21 @@
+
+export const sessionPresenters = {
+  main: {
+    $xsmall: {
+      mB: 8,
+    },
+    $small: {
+      mB: 18,
+    }
+  },
+  text: {
+    $xsmall: {
+      mT: 8,
+      ftSz: 14,
+    },
+    $small: {
+      mT: 20,
+      ftSz: 16,
+    },
+  },
+}
